chore: move store and rename signal to utils
