+++ packages/qwik/src/core/debug.ts
@@ -2,8 +2,8 @@ import { isQrl } from '../server/prefetch-strategy';
 import { isJSXNode } from './shared/jsx/jsx-runtime';
 import { isTask } from './use/use-task';
 import { vnode_getProp, vnode_isVNode } from './client/vnode';
-import { isSignal } from './reactive-primitives/signal';
-import { isStore } from './reactive-primitives/store';
+import { isSignal } from './reactive-primitives/utils';
+import { isStore } from './reactive-primitives/impl/store';
 import { DEBUG_TYPE } from './shared/types';
 import { ComputedSignalImpl } from './reactive-primitives/impl/computed-signal-impl';
 import { WrappedSignalImpl } from './reactive-primitives/impl/wrapped-signal-impl';
