+++ packages/qwik/src/core/reactive-primitives/impl/computed-signal-impl.ts
@@ -5,8 +5,8 @@ import type { Container } from '../../shared/types';
 import { ChoreType } from '../../shared/util-chore-type';
 import { isPromise } from '../../shared/utils/promises';
 import { tryGetInvokeContext } from '../../use/use-core';
-import { throwIfQRLNotResolved } from '../signal';
-import type { BackRef } from '../signal-cleanup';
+import { throwIfQRLNotResolved } from '../utils';
+import type { BackRef } from '../cleanup';
 import { getSubscriber } from '../subscriber';
 import type { ComputeQRL, EffectSubscription } from '../types';
 import { _EFFECT_BACK_REF, EffectProperty, NEEDS_COMPUTATION, SignalFlags } from '../types';
