+++ packages/qwik/src/core/reactive-primitives/signal-api.ts
@@ -2,7 +2,7 @@ import type { QRLInternal } from '../shared/qrl/qrl-class';
 import type { QRL } from '../shared/qrl/qrl.public';
 import { SignalImpl } from './impl/signal-impl';
 import { ComputedSignalImpl } from './impl/computed-signal-impl';
-import { throwIfQRLNotResolved } from './signal';
+import { throwIfQRLNotResolved } from './utils';
 import type { Signal } from './signal.public';
 import type { SerializerArg } from './types';
 import { SerializerSignalImpl } from './impl/serializer-signal-impl';
