+++ packages/qwik/src/core/reactive-primitives/subscriber.ts
@@ -1,7 +1,7 @@
 import { isServer } from '@qwik.dev/core/build';
 import { QBackRefs } from '../shared/utils/markers';
 import type { ISsrNode } from '../ssr/ssr-types';
-import { BackRef } from './signal-cleanup';
+import { BackRef } from './cleanup';
 import type { Consumer, EffectProperty, EffectSubscription } from './types';
 import { _EFFECT_BACK_REF, EffectSubscriptionProp } from './types';
