+++ packages/qwik/src/core/use/use-computed.ts
@@ -2,7 +2,7 @@ import { implicit$FirstArg } from '../shared/qrl/implicit_dollar';
 import { assertQrl } from '../shared/qrl/qrl-utils';
 import type { QRL } from '../shared/qrl/qrl.public';
 import { ComputedSignalImpl } from '../reactive-primitives/impl/computed-signal-impl';
-import { throwIfQRLNotResolved } from '../reactive-primitives/signal';
+import { throwIfQRLNotResolved } from '../reactive-primitives/utils';
 import type { ReadonlySignal, Signal } from '../reactive-primitives/signal.public';
 import { useSequentialScope } from './use-sequential-scope';
